backend/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
# data handling
pandas>=2.2
numpy>=1.26
rapidfuzz>=3.0
